requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
beautifulsoup4==4.12.2
lxml
python-dotenv==1.0.0
apscheduler==3.10.4
requests==2.31.0
aiofiles==23.2.1
python-multipart==0.0.6
aiohttp
